moved .viewAvailableActions and .exportState to separate files
